Move classes up
